add docs, update version
