+++ kotlin/buildSrc/build.gradle.kts
@@ -1,6 +1,9 @@
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
+    // This is the first time kotlin is present during the build
+    // Not providing the version here, makes `kotlin-dsl` use older version
+    // That leads to `jvmTarget` of `17` erroring out
     kotlin("jvm") version "1.7.20"
     `kotlin-dsl`
 }
