+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/Condition.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining results from boolean operations like equals, greater than, etc.
