+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/ConditionEvaluationOptions.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining options passed during evaluation of a condition
