+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/ConditionJoiner.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Enum defining possible ways of joining multiple conditions within a rule
