+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/ConditionResult.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining the possible outputs of a condition evaluation
