+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/EngineEvaluationType.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Enum defining different ways an engine will evaluate the rule set
