+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/EngineOptions.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining properties for a rules engine instance.
