+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/EvaluationResult.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining result of an engine evaluation of the rule set
