+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/ExitCriteria.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining possible exit criteria for an engine evaluation
