+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/Operator.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining an operation type along with the value to perform it against
