+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/Rule.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 import java.time.Instant
