+++ kotlin/core/src/main/kotlin/com/warnermedia/rulesengine/core/RuleEvaluationOptions.kt
@@ -1,4 +1,4 @@
-package core
+package com.warnermedia.rulesengine.core
 
 /**
  * Class defining options passed to a Rule during runtime evaluation
