+++ kotlin/core/src/test/kotlin/com/warnermedia/rulesengine/core/EngineTest.kt
@@ -1,10 +1,9 @@
-package jsonadapter
+package com.warnermedia.rulesengine.core
 
-import core.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-internal class AdapterTest {
+internal class EngineTest {
     private val engine = Engine(
         "weather-type-engine",
         arrayListOf(
@@ -20,12 +19,19 @@ internal class AdapterTest {
     )
 
     @Test
-    fun testPersistence() {
-        Adapter.saveToFile(engine, "engine.json")
-        val engineFromFile = Adapter.readFromFile("engine.json")
-        val result = engineFromFile.evaluate(hashMapOf("temperature" to 75, "rainfall" to 0))
+    fun testSuccessfulResult() {
+        val result = engine.evaluate(hashMapOf("temperature" to 75, "rainfall" to 0))
         assertEquals<ArrayList<RuleResult>>(
             arrayListOf(RuleResult.Success("good-weather", "good-weather-day")),
+            result.ruleEvaluations,
+        )
+    }
+
+    @Test
+    fun testFailureResult() {
+        val result = engine.evaluate(hashMapOf("temperature" to 60, "rainfall" to 0))
+        assertEquals<ArrayList<RuleResult>>(
+            arrayListOf(RuleResult.Failure("good-weather", "work-from-home-day")),
             result.ruleEvaluations,
         )
     }
